# backend/operator-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance analytics and aggregation for operators",
        "title": "Betasafar Operator API (Go)",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.betasafar.app/operator",
    "basePath": "/",
    "paths": {
...
        "/notifications/broadcast": {
            "post": {
                "description": "Operator broadcasts update to all booked pilgrims (placeholder)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Broadcast to package pilgrims",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Package ID",
                        "name": "packageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message title",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Message body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    }
}